docs(eslint-plugin): rewrite member-ordering docs (#349)
Part of the [PR to enforce an alphabetic order for members](https://github.com/typescript-eslint/typescript-eslint/pull/2639) is to rewrite the existing docs to better document the current behavior of `member-ordering`. As this touches almost any line in the rule documentation, I thought it makes sense to open a separate PR for that.

In this PR, I've reworked the entire rule documentation to make it (hopefully) more clear how this rule works. The examples are mostly taken from the current rule documentation - some with little adaptations.
Besides that, I've changed a few names and added some comments in the rule implementation to make it more understandable (basically the points I didn't understand immediately while trying to understand the current implementation to better describe the rule).