fix(eslint-plugin): [space-infix-ops] fix no error when right type is function (#4848)
* fix(eslint-plugin): [space-infix-ops] no error when right type
is function

No errors were thrown when the right type is a function when no
space between union or intersection operator. Function types are
wrapped inside parenthesize. If the right type is function, instead of
getting the operator, it was always getting the function parenthesize.
The fix providedd skips wrapped parenthesize for function type, so that
operator will be properly identified.

* fix: edge case where function has multiple
parenthesis