docs(website): help troubleshoot issues with multiple parsers (#5612)
* Help troubleshoot issues with multiple parsers

This might not be a perfect fit for the typescript-eslint docs, since this is specific to Vue.  However, I expect the issue effects other custom parsers as well. Including multiple parsers like this is counter intuitive because it seems like a mistake to include two "parser"s.  It would have been a huge help to me to have found this in the troubleshooting guide.  

more info:
https://eslint.vuejs.org/user-guide/#how-to-use-a-custom-parser

devs running into this issue:
https://github.com/vuejs/vue-eslint-parser/issues/45#issuecomment-988625895
https://stackoverflow.com/questions/66597732/eslint-vue-3-parsing-error-expected-eslint

* fix typos

* grouping vue related questions

* Apply suggestions from code review

* More syntax highlighting

Co-authored-by: Josh Goldberg <git@joshuakgoldberg.com>