feat(eslint-plugin): [promise-function-async] check for promises in implicit return types (#6330)
* [promise-function-async] Only allow unions in explicit return types

When we return a union containing a promise from a function
implicitly, it's often a mistake. This commit makes it so if the
return type is explicit, any `Promise` in the return type (whether
it's part of a union or not) will flag the function. If it is
intentional, make the return type explicit.

Fixes #6329

* Refrain from renaming the type-util, instead add an optional fourth param.